Tugas 9

+++ counter_7/lib/model/watchlist.dart
@@ -50,7 +50,7 @@ class Fields {
   factory Fields.fromJson(Map<String, dynamic> json) => Fields(
     watched: json["watched"],
     title: json["title"],
-    rating: json["rating"].toDouble(),
+    rating: json["rating"],
     releaseDate: json["release_date"],
     review: json["review"],
   );
